+++ docs/pages/components/upload/examples/ExSimple.vue
@@ -1,16 +1,15 @@
 <template>
-    <b-field>
+    <b-field class="file">
         <b-upload v-model="files">
             <a class="button is-primary">
                 <b-icon icon="upload"></b-icon>
                 <span>Click to upload</span>
             </a>
         </b-upload>
-        <div v-if="files && files.length">
-            <span class="file-name">
-                {{ files[0].name }}
-            </span>
-        </div>
+        <span class="file-name"
+            v-if="files && files.length">
+            {{ files[0].name }}
+        </span>
     </b-field>
 </template>